+++ incs/minishell.h
@@ -6,7 +6,7 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/25 10:08:39 by marvin            #+#    #+#             */
-/*   Updated: 2023/09/24 12:25:04 by codespace        ###   ########.fr       */
+/*   Updated: 2023/09/24 13:01:33 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -202,11 +202,14 @@ int		worker_task_preparation(t_block *worker);
 int		worker_extract_redirections(t_block *worker);
 
 /*worker_tasks_2.c*/
-int 	worker_args_split_add_token(t_block *worker, t_token_node *arg, int *move);
 int		worker_args_rm_unguarded_quotes(t_block *worker);
 int		worker_args_split_unguarded_quotes(t_block *worker);
 int		worker_args_expand_dollar_wildcard(t_block *worker);
 int		worker_dump_tasks_to_cmd_args(t_block *worker);
+
+/*worker_tasks_3.c*/
+int 	worker_args_split_add_token(t_block *worker, t_token_node *arg, int *move);
+
 
 //////////////////////////////////////
 //////////////////////////////////////
@@ -305,7 +308,7 @@ int		open_here_docs_at_block(t_block *block);
 //////////////////////////////////////
 
 /*wildcard_expansion.c*/
-int		expand_wildcards(char **to_expand, char **fail_return);
+int		expand_wildcards(char **to_expand);
 
 /*wildcard_return.c*/
 char	*wildcard(char *pattern, int pat_len, int *match_count);
